+++ test/form.test.ts
@@ -730,6 +730,29 @@ test("setErrors", async () => {
   expect(field.error).toEqual("WRONG");
 });
 
+test("setErrors repeating", async () => {
+  const N = types.model("N", {
+    bar: types.string
+  });
+  const M = types.model("M", {
+    foo: types.array(N)
+  });
+
+  const form = new Form(M, {
+    foo: new RepeatingForm({
+      bar: new Field(converters.string)
+    })
+  });
+
+  const o = M.create({ foo: [{ bar: "FOOO" }] });
+
+  const state = form.state(o);
+  state.setErrors({ foo: [{ bar: "WRONG" }] });
+
+  const field = state.repeatingForm("foo").accessors[0].field("bar");
+  expect(field.error).toBe("WRONG");
+});
+
 test("additional error by name", async () => {
   const M = types.model("M", {
     foo: types.string